Implemented partial config parser/generator
Major refactoring for uniformity/cleanliness
Implmented partial UI frontend logic for bladeidpage.

+++ src/sources/progress.cpp
@@ -7,7 +7,7 @@
 
 wxEventTypeTag<wxCommandEvent> Progress::EVT_UPDATE(wxNewEventType());
 
-void Progress::emitEvent(int progress, wxString message) {
+void Progress::emitEvent(int8_t progress, wxString message) {
 
   ProgressEvent* event = new ProgressEvent(EVT_UPDATE, wxID_ANY);
   event->progress = progress;
